% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{betas.post}
\alias{betas.post}
\title{betas.post}
\usage{
betas.post(betas.p, p, Sigma.b, Y, Y0, X, W, betas, delta, C, rho, form)
}
\arguments{
\item{betas.p}{current value of the pth element of betas}

\item{p}{pth element}

\item{Sigma.b}{variance estimate of betas}

\item{Y}{response variable}

\item{Y0}{add...}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of betas with pth element fixed as betas.p
}
