% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{betas.post2}
\alias{betas.post2}
\title{betas.post2}
\usage{
betas.post2(betas.p, p, Sigma.b, Y, Y0, X, betas, alpha, C, lambda, form)
}
\arguments{
\item{betas.p}{current value of the pth element of betas}

\item{p}{pth element}

\item{Sigma.b}{variance estimate of betas}

\item{Y}{the time (duration) dependent variable for the survival stage (t).}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of betas with pth element fixed as betas.p
}
