% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{betas.slice.sampling2}
\alias{betas.slice.sampling2}
\title{betas.slice.sampling2}
\usage{
betas.slice.sampling2(Sigma.b, Y, Y0, X, betas, alpha, C, lambda, w, m, form)
}
\arguments{
\item{Sigma.b}{variance estimate of betas}

\item{Y}{the time (duration) dependent variable for the survival stage (t).}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for betas
}
