% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{gammas.post}
\alias{gammas.post}
\title{gammas.post}
\usage{
gammas.post(gammas.p, p, Sigma.g, Y, eXB, Z, gammas, C, rho, form)
}
\arguments{
\item{gammas.p}{current value of the pth element of gammas}

\item{p}{pth element}

\item{Sigma.g}{variance estimate of gammas}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of gammas with pth element fixed as gammas.p
}
