% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{gammas.slice.sampling}
\alias{gammas.slice.sampling}
\title{gammas.slice.sampling}
\usage{
gammas.slice.sampling(Sigma.g, Y, eXB, Z, gammas, C, rho, w, m, form)
}
\arguments{
\item{Sigma.g}{variance estimate of gammas}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for gammas
}
