% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{gammas.slice.sampling2}
\alias{gammas.slice.sampling2}
\title{gammas.slice.sampling2}
\usage{
gammas.slice.sampling2(Sigma.g, Y, Y0, eXB, Z, gammas, C, lambda, w, m, form)
}
\arguments{
\item{Sigma.g}{variance estimate of gammas}

\item{Y}{the time (duration) dependent variable for the survival stage (t).}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for gammas
}
