% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{lambda.post2}
\alias{lambda.post2}
\title{lambda.post2}
\usage{
lambda.post2(Y, Y0, eXB, alpha, C, lambda, a = 1, b = 1)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t).}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{eXB}{exponentiated vector of covariates times betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of lambda
}
