% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl2.R
\name{lambda.slice.sampling2}
\alias{lambda.slice.sampling2}
\title{lambda.slice.sampling2}
\usage{
lambda.slice.sampling2(
  Y,
  Y0,
  eXB,
  alpha,
  C,
  lambda,
  w,
  m,
  lower = 0.01,
  upper = +Inf
)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t).}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{eXB}{exponentiated vector of covariates times betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for lambda
}
