% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{mcmcSpatialCure2}
\alias{mcmcSpatialCure2}
\title{mcmcSpatialCure2}
\usage{
mcmcSpatialCure2(
  Y,
  C,
  X,
  Z,
  S,
  A,
  N,
  burn,
  thin,
  w = c(1, 1, 1),
  m = 10,
  form,
  prop.var
)
}
\arguments{
\item{Y}{response variable}

\item{C}{censoring indicator}

\item{X}{covariates for betas}

\item{Z}{covariates for gammas}

\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho)}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
chain of the variables of interest
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian spatial cure model
}
