% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcsurv.R
\name{mcmcsurv}
\alias{mcmcsurv}
\alias{summary.mcmcsurv}
\title{mcmcsurv}
\usage{
mcmcsurv(Y, Y0, C, X, N, burn, thin, w = c(1, 1, 1), m = 10, form)

\method{summary}{mcmcsurv}(object, parameter = c("betas", "rho"), ...)
}
\arguments{
\item{Y}{response variable}

\item{Y0}{add...}

\item{C}{censoring indicator}

\item{X}{covariates for betas}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho)}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}

\item{object}{an object of class \code{mfsurv}, the output of \code{\link{mfsurv}}.}

\item{parameter}{one of three parameters of the mfsurv output. Indicate either "betas" or "rho".}

\item{...}{additional parameter}
}
\value{
chain of the variables of interest

list. Empirical mean, standard deviation and quantiles for each variable.
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian survival (Weibull) model

Returns a summary of a mfsurv object via \code{\link[coda]{summary.mcmc}}.
}
