% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.mfsurv.R
\name{mfsurv.stats}
\alias{mfsurv.stats}
\title{mfsurv.stats}
\usage{
mfsurv.stats(object)
}
\arguments{
\item{object}{an object of class \code{mfsurv}, the output of \code{mfsurv()}.}
}
\description{
A function to calculate the deviance information criterion (DIC) for fitted model objects of class \code{mfsurv}
for which a log-likelihood can be obtained, according to the formula \emph{DIC = -2 * (L - P)},
where \emph{L} is the log likelihood of the data given the posterior means of the parameter and
\emph{P} is the  estimate of the effective number of parameters in the model.
}
\examples{
\dontrun{
model1 <- mfsurv(Y ~ X1 + X2 | C ~ Z1 + Z2, Y0 = Y0, N = 1000, burn = 50, thin = 5,
                 w = c(0.1, .1, .1), m = 10, data = data, form = "Weibull",
                 na.action = 'na.omit')

mfsurv.stats(model1)
}

}
