% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.mfsurv.R
\name{mfsurv.summary}
\alias{mfsurv.summary}
\title{mfsurv.summary()}
\usage{
mfsurv.summary(object, parameter = c("betas", "gammas", "lambda"))
}
\arguments{
\item{object}{an object of class \code{mfsurv}, the output of \code{\link{mfsurv}}.}

\item{parameter}{one of three parameters of the mfsurv output. Indicate either "betas", "gammas" or "lambda".}
}
\description{
Returns a summary of a mfsurv object via \code{\link[coda]{summary.mcmc}}.
}
\examples{
\dontrun{
model1 <- mfsurv(Y ~ X1 + X2 | C ~ Z1 + Z2, Y0 = Y0, N = 1000, burn = 50, thin = 5,
                 w = c(0.1, .1, .1), m = 10, data = data, form = "Weibull",
                 a.action = 'na.omit')

mfsurv.summary(model1, "betas")
}
}
\seealso{
\code{\link{mcmc}}, \code{\link{summary.mcmc}}
}
