% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{rho.post}
\alias{rho.post}
\title{rho.post}
\usage{
rho.post(Y, Y0, eXB, delta, C, rho, a = 1, b = 1)
}
\arguments{
\item{Y}{response variable}

\item{Y0}{add...}

\item{eXB}{exponentiated vector of covariates times betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of rho
}
