% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{rho.slice.sampling}
\alias{rho.slice.sampling}
\title{rho.slice.sampling}
\usage{
rho.slice.sampling(Y, Y0, eXB, delta, C, rho, w, m, lower = 0.01, upper = +Inf)
}
\arguments{
\item{Y}{response variable}

\item{Y0}{add...}

\item{eXB}{exponentiated vector of covariates times betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for rho
}
