% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{V.MH.sampling}
\alias{V.MH.sampling}
\title{V.MH.sampling}
\usage{
V.MH.sampling(S, A, lambda, Y, eXB, Z, V, gammas, C, rho, prop.var)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for W
}
