% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{V.post}
\alias{V.post}
\title{V.post}
\usage{
V.post(S, A, lambda, Y, eXB, Z, V, gammas, C, rho)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of W with sth element fixed as W.s
}
