% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgl.R
\name{W.MH.sampling}
\alias{W.MH.sampling}
\title{W.MH.sampling}
\usage{
W.MH.sampling(S, A, lambda, Y, X, W, betas, delta, C, rho, prop.var)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for W
}
