% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSurv.R
\name{mcmcsurv}
\alias{mcmcsurv}
\alias{summary.mcmcsurv}
\title{mcmcsurv}
\usage{
mcmcsurv(Y, Y0, C, X, N, burn, thin, w = c(1, 1, 1), m = 10, form)

\method{summary}{mcmcsurv}(object, parameter = c("betas", "rho"), ...)
}
\arguments{
\item{Y}{response variable.}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1).}

\item{C}{censoring indicator.}

\item{X}{covariates for betas.}

\item{N}{number of MCMC iterations.}

\item{burn}{burn-in to be discarded.}

\item{thin}{thinning to prevent from autocorrelation.}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho).}

\item{m}{limit on steps in the slice sampling.}

\item{form}{type of parametric model (Exponential or Weibull).}

\item{object}{an object of class \code{mfsurv}, the output of \code{\link{mfsurv}}.}

\item{parameter}{one of three parameters of the mfsurv output. Indicate either "betas" or "rho".}

\item{...}{additional parameter}
}
\value{
chain of the variables of interest.

list. Empirical mean, standard deviation and quantiles for each variable.
}
\description{
\code{mcmcsurv} estimates a Bayesian Exponential or Weibull survival model via Markov Chain Monte Carlo (MCMC). Slice samplig is employed to draw the posterior sample of the model's survival stage parameters.

Returns a summary of a mfsurv object via \code{\link[coda]{summary.mcmc}}.
}
\examples{
set.seed(95)
bgl <- Buhaugetal_2009_JCR
bgl <- subset(bgl, coupx == 0)
bgl <- na.omit(bgl)
Y   <- bgl$Y
X   <- as.matrix(cbind(1, bgl[,1:7]))
C   <- bgl$C
Y0  <- bgl$Y0
model2 <- mcmcsurv(Y = Y, Y0 = Y0, C =  C,  X = X,
                   N = 50,
                   burn = 20,
                   thin = 15,
                   w = c(0.5, 0.5, 0.5),
                   m = 5,
                   form = "Weibull")

summary(model2, parameter = "betas")

}
