% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSurv.R
\name{mcmcSurv}
\alias{mcmcSurv}
\title{mcmcSurv}
\usage{
mcmcSurv(Y, Y0, C, X, N, burn, thin, w = c(1, 1, 1), m = 10, form)
}
\arguments{
\item{Y}{response variable}

\item{Y0}{add...}

\item{C}{censoring indicator}

\item{X}{covariates for betas}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho)}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
chain of the variables of interest
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian survival (Weibull) model
}
